+++ lib/src/websocket_request_manager/interfaces/socket_response.dart
@@ -1,0 +1,10 @@
+import '../../../websocket_universal.dart';
+
+/// Grouped reponse data to `ISocketRequest`
+abstract class ISocketResponse {
+  /// original request
+  ISocketMessage get requestMessage;
+
+  /// aggregated responses
+  List<ISocketMessage> get responses;
+}