+++ lib/src/websocket_request_manager/interfaces/timed_socket_response.dart
@@ -1,0 +1,7 @@
+import '../../../websocket_universal.dart';
+
+/// [ISocketMessage] with timestamp
+abstract class ITimedSocketResponse<T> implements ISocketMessage<T> {
+  /// When message was received
+  DateTime get timestamp;
+}